Move storage for Mesh, MeshInstance, MultiMesh and Skeleton into MeshStorage
