+++ servers/rendering/renderer_rd/renderer_storage_rd.h
@@ -39,12 +39,8 @@
 #include "servers/rendering/renderer_rd/shaders/canvas_sdf.glsl.gen.h"
 #include "servers/rendering/renderer_rd/shaders/particles.glsl.gen.h"
 #include "servers/rendering/renderer_rd/shaders/particles_copy.glsl.gen.h"
-#include "servers/rendering/renderer_rd/shaders/skeleton.glsl.gen.h"
 #include "servers/rendering/renderer_rd/shaders/voxel_gi_sdf.glsl.gen.h"
-#include "servers/rendering/renderer_rd/storage_rd/canvas_texture_storage.h"
-#include "servers/rendering/renderer_rd/storage_rd/decal_atlas_storage.h"
 #include "servers/rendering/renderer_rd/storage_rd/material_storage.h"
-#include "servers/rendering/renderer_rd/storage_rd/texture_storage.h"
 #include "servers/rendering/renderer_scene_render.h"
 #include "servers/rendering/rendering_device.h"
 #include "servers/rendering/shader_compiler.h"
@@ -129,232 +125,11 @@ public:
 		}
 	}
 
-	enum DefaultRDBuffer {
-		DEFAULT_RD_BUFFER_VERTEX,
-		DEFAULT_RD_BUFFER_NORMAL,
-		DEFAULT_RD_BUFFER_TANGENT,
-		DEFAULT_RD_BUFFER_COLOR,
-		DEFAULT_RD_BUFFER_TEX_UV,
-		DEFAULT_RD_BUFFER_TEX_UV2,
-		DEFAULT_RD_BUFFER_CUSTOM0,
-		DEFAULT_RD_BUFFER_CUSTOM1,
-		DEFAULT_RD_BUFFER_CUSTOM2,
-		DEFAULT_RD_BUFFER_CUSTOM3,
-		DEFAULT_RD_BUFFER_BONES,
-		DEFAULT_RD_BUFFER_WEIGHTS,
-		DEFAULT_RD_BUFFER_MAX,
-	};
-
 private:
 	/* TEXTURE API */
 
 	RID default_rd_samplers[RS::CANVAS_ITEM_TEXTURE_FILTER_MAX][RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX];
 	RID custom_rd_samplers[RS::CANVAS_ITEM_TEXTURE_FILTER_MAX][RS::CANVAS_ITEM_TEXTURE_REPEAT_MAX];
-	RID default_rd_storage_buffer;
-
-	/* Mesh */
-
-	struct MeshInstance;
-
-	struct Mesh {
-		struct Surface {
-			RS::PrimitiveType primitive = RS::PRIMITIVE_POINTS;
-			uint32_t format = 0;
-
-			RID vertex_buffer;
-			RID attribute_buffer;
-			RID skin_buffer;
-			uint32_t vertex_count = 0;
-			uint32_t vertex_buffer_size = 0;
-			uint32_t skin_buffer_size = 0;
-
-			// A different pipeline needs to be allocated
-			// depending on the inputs available in the
-			// material.
-			// There are never that many geometry/material
-			// combinations, so a simple array is the most
-			// cache-efficient structure.
-
-			struct Version {
-				uint32_t input_mask = 0;
-				RD::VertexFormatID vertex_format = 0;
-				RID vertex_array;
-			};
-
-			SpinLock version_lock; //needed to access versions
-			Version *versions = nullptr; //allocated on demand
-			uint32_t version_count = 0;
-
-			RID index_buffer;
-			RID index_array;
-			uint32_t index_count = 0;
-
-			struct LOD {
-				float edge_length = 0.0;
-				uint32_t index_count = 0;
-				RID index_buffer;
-				RID index_array;
-			};
-
-			LOD *lods = nullptr;
-			uint32_t lod_count = 0;
-
-			AABB aabb;
-
-			Vector<AABB> bone_aabbs;
-
-			RID blend_shape_buffer;
-
-			RID material;
-
-			uint32_t render_index = 0;
-			uint64_t render_pass = 0;
-
-			uint32_t multimesh_render_index = 0;
-			uint64_t multimesh_render_pass = 0;
-
-			uint32_t particles_render_index = 0;
-			uint64_t particles_render_pass = 0;
-
-			RID uniform_set;
-		};
-
-		uint32_t blend_shape_count = 0;
-		RS::BlendShapeMode blend_shape_mode = RS::BLEND_SHAPE_MODE_NORMALIZED;
-
-		Surface **surfaces = nullptr;
-		uint32_t surface_count = 0;
-
-		Vector<AABB> bone_aabbs;
-
-		bool has_bone_weights = false;
-
-		AABB aabb;
-		AABB custom_aabb;
-
-		Vector<RID> material_cache;
-
-		List<MeshInstance *> instances;
-
-		RID shadow_mesh;
-		Set<Mesh *> shadow_owners;
-
-		Dependency dependency;
-	};
-
-	mutable RID_Owner<Mesh, true> mesh_owner;
-
-	struct MeshInstance {
-		Mesh *mesh;
-		RID skeleton;
-		struct Surface {
-			RID vertex_buffer;
-			RID uniform_set;
-
-			Mesh::Surface::Version *versions = nullptr; //allocated on demand
-			uint32_t version_count = 0;
-		};
-		LocalVector<Surface> surfaces;
-		LocalVector<float> blend_weights;
-
-		RID blend_weights_buffer;
-		List<MeshInstance *>::Element *I = nullptr; //used to erase itself
-		uint64_t skeleton_version = 0;
-		bool dirty = false;
-		bool weights_dirty = false;
-		SelfList<MeshInstance> weight_update_list;
-		SelfList<MeshInstance> array_update_list;
-		MeshInstance() :
-				weight_update_list(this), array_update_list(this) {}
-	};
-
-	void _mesh_instance_clear(MeshInstance *mi);
-	void _mesh_instance_add_surface(MeshInstance *mi, Mesh *mesh, uint32_t p_surface);
-
-	mutable RID_Owner<MeshInstance> mesh_instance_owner;
-
-	SelfList<MeshInstance>::List dirty_mesh_instance_weights;
-	SelfList<MeshInstance>::List dirty_mesh_instance_arrays;
-
-	struct SkeletonShader {
-		struct PushConstant {
-			uint32_t has_normal;
-			uint32_t has_tangent;
-			uint32_t has_skeleton;
-			uint32_t has_blend_shape;
-
-			uint32_t vertex_count;
-			uint32_t vertex_stride;
-			uint32_t skin_stride;
-			uint32_t skin_weight_offset;
-
-			uint32_t blend_shape_count;
-			uint32_t normalized_blend_shapes;
-			uint32_t pad0;
-			uint32_t pad1;
-		};
-
-		enum {
-			UNIFORM_SET_INSTANCE = 0,
-			UNIFORM_SET_SURFACE = 1,
-			UNIFORM_SET_SKELETON = 2,
-		};
-		enum {
-			SHADER_MODE_2D,
-			SHADER_MODE_3D,
-			SHADER_MODE_MAX
-		};
-
-		SkeletonShaderRD shader;
-		RID version;
-		RID version_shader[SHADER_MODE_MAX];
-		RID pipeline[SHADER_MODE_MAX];
-
-		RID default_skeleton_uniform_set;
-	} skeleton_shader;
-
-	void _mesh_surface_generate_version_for_input_mask(Mesh::Surface::Version &v, Mesh::Surface *s, uint32_t p_input_mask, MeshInstance::Surface *mis = nullptr);
-
-	RID mesh_default_rd_buffers[DEFAULT_RD_BUFFER_MAX];
-
-	/* MultiMesh */
-	struct MultiMesh {
-		RID mesh;
-		int instances = 0;
-		RS::MultimeshTransformFormat xform_format = RS::MULTIMESH_TRANSFORM_3D;
-		bool uses_colors = false;
-		bool uses_custom_data = false;
-		int visible_instances = -1;
-		AABB aabb;
-		bool aabb_dirty = false;
-		bool buffer_set = false;
-		uint32_t stride_cache = 0;
-		uint32_t color_offset_cache = 0;
-		uint32_t custom_data_offset_cache = 0;
-
-		Vector<float> data_cache; //used if individual setting is used
-		bool *data_cache_dirty_regions = nullptr;
-		uint32_t data_cache_used_dirty_regions = 0;
-
-		RID buffer; //storage buffer
-		RID uniform_set_3d;
-		RID uniform_set_2d;
-
-		bool dirty = false;
-		MultiMesh *dirty_list = nullptr;
-
-		Dependency dependency;
-	};
-
-	mutable RID_Owner<MultiMesh, true> multimesh_owner;
-
-	MultiMesh *multimesh_dirty_list = nullptr;
-
-	_FORCE_INLINE_ void _multimesh_make_local(MultiMesh *multimesh) const;
-	_FORCE_INLINE_ void _multimesh_mark_dirty(MultiMesh *multimesh, int p_index, bool p_aabb);
-	_FORCE_INLINE_ void _multimesh_mark_all_dirty(MultiMesh *multimesh, bool p_data, bool p_aabb);
-	_FORCE_INLINE_ void _multimesh_re_create_aabb(MultiMesh *multimesh, const float *p_data, int p_instances);
-	void _update_dirty_multimeshes();
 
 	/* PARTICLES */
 
@@ -732,34 +507,6 @@ private:
 
 	mutable RID_Owner<VisibilityNotifier> visibility_notifier_owner;
 
-	/* Skeleton */
-
-	struct Skeleton {
-		bool use_2d = false;
-		int size = 0;
-		Vector<float> data;
-		RID buffer;
-
-		bool dirty = false;
-		Skeleton *dirty_list = nullptr;
-		Transform2D base_transform_2d;
-
-		RID uniform_set_3d;
-		RID uniform_set_mi;
-
-		uint64_t version = 1;
-
-		Dependency dependency;
-	};
-
-	mutable RID_Owner<Skeleton, true> skeleton_owner;
-
-	_FORCE_INLINE_ void _skeleton_make_dirty(Skeleton *skeleton);
-
-	Skeleton *skeleton_dirty_list = nullptr;
-
-	void _update_dirty_skeletons();
-
 	/* LIGHT */
 
 	struct Light {
@@ -975,360 +722,6 @@ public:
 
 	void sampler_rd_set_default(float p_mipmap_bias);
 
-	/* MESH API */
-
-	RID mesh_allocate();
-	void mesh_initialize(RID p_mesh);
-
-	virtual void mesh_set_blend_shape_count(RID p_mesh, int p_blend_shape_count);
-
-	/// Return stride
-	virtual void mesh_add_surface(RID p_mesh, const RS::SurfaceData &p_surface);
-
-	virtual int mesh_get_blend_shape_count(RID p_mesh) const;
-
-	virtual void mesh_set_blend_shape_mode(RID p_mesh, RS::BlendShapeMode p_mode);
-	virtual RS::BlendShapeMode mesh_get_blend_shape_mode(RID p_mesh) const;
-
-	virtual void mesh_surface_update_vertex_region(RID p_mesh, int p_surface, int p_offset, const Vector<uint8_t> &p_data);
-	virtual void mesh_surface_update_attribute_region(RID p_mesh, int p_surface, int p_offset, const Vector<uint8_t> &p_data);
-	virtual void mesh_surface_update_skin_region(RID p_mesh, int p_surface, int p_offset, const Vector<uint8_t> &p_data);
-
-	virtual void mesh_surface_set_material(RID p_mesh, int p_surface, RID p_material);
-	virtual RID mesh_surface_get_material(RID p_mesh, int p_surface) const;
-
-	virtual RS::SurfaceData mesh_get_surface(RID p_mesh, int p_surface) const;
-
-	virtual int mesh_get_surface_count(RID p_mesh) const;
-
-	virtual void mesh_set_custom_aabb(RID p_mesh, const AABB &p_aabb);
-	virtual AABB mesh_get_custom_aabb(RID p_mesh) const;
-
-	virtual AABB mesh_get_aabb(RID p_mesh, RID p_skeleton = RID());
-	virtual void mesh_set_shadow_mesh(RID p_mesh, RID p_shadow_mesh);
-
-	virtual void mesh_clear(RID p_mesh);
-
-	virtual bool mesh_needs_instance(RID p_mesh, bool p_has_skeleton);
-
-	/* MESH INSTANCE */
-
-	virtual RID mesh_instance_create(RID p_base);
-	virtual void mesh_instance_set_skeleton(RID p_mesh_instance, RID p_skeleton);
-	virtual void mesh_instance_set_blend_shape_weight(RID p_mesh_instance, int p_shape, float p_weight);
-	virtual void mesh_instance_check_for_update(RID p_mesh_instance);
-	virtual void update_mesh_instances();
-
-	_FORCE_INLINE_ const RID *mesh_get_surface_count_and_materials(RID p_mesh, uint32_t &r_surface_count) {
-		Mesh *mesh = mesh_owner.get_or_null(p_mesh);
-		ERR_FAIL_COND_V(!mesh, nullptr);
-		r_surface_count = mesh->surface_count;
-		if (r_surface_count == 0) {
-			return nullptr;
-		}
-		if (mesh->material_cache.is_empty()) {
-			mesh->material_cache.resize(mesh->surface_count);
-			for (uint32_t i = 0; i < r_surface_count; i++) {
-				mesh->material_cache.write[i] = mesh->surfaces[i]->material;
-			}
-		}
-
-		return mesh->material_cache.ptr();
-	}
-
-	_FORCE_INLINE_ void *mesh_get_surface(RID p_mesh, uint32_t p_surface_index) {
-		Mesh *mesh = mesh_owner.get_or_null(p_mesh);
-		ERR_FAIL_COND_V(!mesh, nullptr);
-		ERR_FAIL_UNSIGNED_INDEX_V(p_surface_index, mesh->surface_count, nullptr);
-
-		return mesh->surfaces[p_surface_index];
-	}
-
-	_FORCE_INLINE_ RID mesh_get_shadow_mesh(RID p_mesh) {
-		Mesh *mesh = mesh_owner.get_or_null(p_mesh);
-		ERR_FAIL_COND_V(!mesh, RID());
-
-		return mesh->shadow_mesh;
-	}
-
-	_FORCE_INLINE_ RS::PrimitiveType mesh_surface_get_primitive(void *p_surface) {
-		Mesh::Surface *surface = reinterpret_cast<Mesh::Surface *>(p_surface);
-		return surface->primitive;
-	}
-
-	_FORCE_INLINE_ bool mesh_surface_has_lod(void *p_surface) const {
-		Mesh::Surface *s = reinterpret_cast<Mesh::Surface *>(p_surface);
-		return s->lod_count > 0;
-	}
-
-	_FORCE_INLINE_ uint32_t mesh_surface_get_vertices_drawn_count(void *p_surface) const {
-		Mesh::Surface *s = reinterpret_cast<Mesh::Surface *>(p_surface);
-		return s->index_count ? s->index_count : s->vertex_count;
-	}
-
-	_FORCE_INLINE_ uint32_t mesh_surface_get_lod(void *p_surface, float p_model_scale, float p_distance_threshold, float p_mesh_lod_threshold, uint32_t *r_index_count = nullptr) const {
-		Mesh::Surface *s = reinterpret_cast<Mesh::Surface *>(p_surface);
-
-		int32_t current_lod = -1;
-		if (r_index_count) {
-			*r_index_count = s->index_count;
-		}
-		for (uint32_t i = 0; i < s->lod_count; i++) {
-			float screen_size = s->lods[i].edge_length * p_model_scale / p_distance_threshold;
-			if (screen_size > p_mesh_lod_threshold) {
-				break;
-			}
-			current_lod = i;
-		}
-		if (current_lod == -1) {
-			return 0;
-		} else {
-			if (r_index_count) {
-				*r_index_count = s->lods[current_lod].index_count;
-			}
-			return current_lod + 1;
-		}
-	}
-
-	_FORCE_INLINE_ RID mesh_surface_get_index_array(void *p_surface, uint32_t p_lod) const {
-		Mesh::Surface *s = reinterpret_cast<Mesh::Surface *>(p_surface);
-
-		if (p_lod == 0) {
-			return s->index_array;
-		} else {
-			return s->lods[p_lod - 1].index_array;
-		}
-	}
-
-	_FORCE_INLINE_ void mesh_surface_get_vertex_arrays_and_format(void *p_surface, uint32_t p_input_mask, RID &r_vertex_array_rd, RD::VertexFormatID &r_vertex_format) {
-		Mesh::Surface *s = reinterpret_cast<Mesh::Surface *>(p_surface);
-
-		s->version_lock.lock();
-
-		//there will never be more than, at much, 3 or 4 versions, so iterating is the fastest way
-
-		for (uint32_t i = 0; i < s->version_count; i++) {
-			if (s->versions[i].input_mask != p_input_mask) {
-				continue;
-			}
-			//we have this version, hooray
-			r_vertex_format = s->versions[i].vertex_format;
-			r_vertex_array_rd = s->versions[i].vertex_array;
-			s->version_lock.unlock();
-			return;
-		}
-
-		uint32_t version = s->version_count;
-		s->version_count++;
-		s->versions = (Mesh::Surface::Version *)memrealloc(s->versions, sizeof(Mesh::Surface::Version) * s->version_count);
-
-		_mesh_surface_generate_version_for_input_mask(s->versions[version], s, p_input_mask);
-
-		r_vertex_format = s->versions[version].vertex_format;
-		r_vertex_array_rd = s->versions[version].vertex_array;
-
-		s->version_lock.unlock();
-	}
-
-	_FORCE_INLINE_ void mesh_instance_surface_get_vertex_arrays_and_format(RID p_mesh_instance, uint32_t p_surface_index, uint32_t p_input_mask, RID &r_vertex_array_rd, RD::VertexFormatID &r_vertex_format) {
-		MeshInstance *mi = mesh_instance_owner.get_or_null(p_mesh_instance);
-		ERR_FAIL_COND(!mi);
-		Mesh *mesh = mi->mesh;
-		ERR_FAIL_UNSIGNED_INDEX(p_surface_index, mesh->surface_count);
-
-		MeshInstance::Surface *mis = &mi->surfaces[p_surface_index];
-		Mesh::Surface *s = mesh->surfaces[p_surface_index];
-
-		s->version_lock.lock();
-
-		//there will never be more than, at much, 3 or 4 versions, so iterating is the fastest way
-
-		for (uint32_t i = 0; i < mis->version_count; i++) {
-			if (mis->versions[i].input_mask != p_input_mask) {
-				continue;
-			}
-			//we have this version, hooray
-			r_vertex_format = mis->versions[i].vertex_format;
-			r_vertex_array_rd = mis->versions[i].vertex_array;
-			s->version_lock.unlock();
-			return;
-		}
-
-		uint32_t version = mis->version_count;
-		mis->version_count++;
-		mis->versions = (Mesh::Surface::Version *)memrealloc(mis->versions, sizeof(Mesh::Surface::Version) * mis->version_count);
-
-		_mesh_surface_generate_version_for_input_mask(mis->versions[version], s, p_input_mask, mis);
-
-		r_vertex_format = mis->versions[version].vertex_format;
-		r_vertex_array_rd = mis->versions[version].vertex_array;
-
-		s->version_lock.unlock();
-	}
-
-	_FORCE_INLINE_ RID mesh_get_default_rd_buffer(DefaultRDBuffer p_buffer) {
-		ERR_FAIL_INDEX_V(p_buffer, DEFAULT_RD_BUFFER_MAX, RID());
-		return mesh_default_rd_buffers[p_buffer];
-	}
-
-	_FORCE_INLINE_ uint32_t mesh_surface_get_render_pass_index(RID p_mesh, uint32_t p_surface_index, uint64_t p_render_pass, uint32_t *r_index) {
-		Mesh *mesh = mesh_owner.get_or_null(p_mesh);
-		Mesh::Surface *s = mesh->surfaces[p_surface_index];
-
-		if (s->render_pass != p_render_pass) {
-			(*r_index)++;
-			s->render_pass = p_render_pass;
-			s->render_index = *r_index;
-		}
-
-		return s->render_index;
-	}
-
-	_FORCE_INLINE_ uint32_t mesh_surface_get_multimesh_render_pass_index(RID p_mesh, uint32_t p_surface_index, uint64_t p_render_pass, uint32_t *r_index) {
-		Mesh *mesh = mesh_owner.get_or_null(p_mesh);
-		Mesh::Surface *s = mesh->surfaces[p_surface_index];
-
-		if (s->multimesh_render_pass != p_render_pass) {
-			(*r_index)++;
-			s->multimesh_render_pass = p_render_pass;
-			s->multimesh_render_index = *r_index;
-		}
-
-		return s->multimesh_render_index;
-	}
-
-	_FORCE_INLINE_ uint32_t mesh_surface_get_particles_render_pass_index(RID p_mesh, uint32_t p_surface_index, uint64_t p_render_pass, uint32_t *r_index) {
-		Mesh *mesh = mesh_owner.get_or_null(p_mesh);
-		Mesh::Surface *s = mesh->surfaces[p_surface_index];
-
-		if (s->particles_render_pass != p_render_pass) {
-			(*r_index)++;
-			s->particles_render_pass = p_render_pass;
-			s->particles_render_index = *r_index;
-		}
-
-		return s->particles_render_index;
-	}
-
-	/* MULTIMESH API */
-
-	RID multimesh_allocate();
-	void multimesh_initialize(RID p_multimesh);
-
-	void multimesh_allocate_data(RID p_multimesh, int p_instances, RS::MultimeshTransformFormat p_transform_format, bool p_use_colors = false, bool p_use_custom_data = false);
-	int multimesh_get_instance_count(RID p_multimesh) const;
-
-	void multimesh_set_mesh(RID p_multimesh, RID p_mesh);
-	void multimesh_instance_set_transform(RID p_multimesh, int p_index, const Transform3D &p_transform);
-	void multimesh_instance_set_transform_2d(RID p_multimesh, int p_index, const Transform2D &p_transform);
-	void multimesh_instance_set_color(RID p_multimesh, int p_index, const Color &p_color);
-	void multimesh_instance_set_custom_data(RID p_multimesh, int p_index, const Color &p_color);
-
-	RID multimesh_get_mesh(RID p_multimesh) const;
-
-	Transform3D multimesh_instance_get_transform(RID p_multimesh, int p_index) const;
-	Transform2D multimesh_instance_get_transform_2d(RID p_multimesh, int p_index) const;
-	Color multimesh_instance_get_color(RID p_multimesh, int p_index) const;
-	Color multimesh_instance_get_custom_data(RID p_multimesh, int p_index) const;
-
-	void multimesh_set_buffer(RID p_multimesh, const Vector<float> &p_buffer);
-	Vector<float> multimesh_get_buffer(RID p_multimesh) const;
-
-	void multimesh_set_visible_instances(RID p_multimesh, int p_visible);
-	int multimesh_get_visible_instances(RID p_multimesh) const;
-
-	AABB multimesh_get_aabb(RID p_multimesh) const;
-
-	_FORCE_INLINE_ RS::MultimeshTransformFormat multimesh_get_transform_format(RID p_multimesh) const {
-		MultiMesh *multimesh = multimesh_owner.get_or_null(p_multimesh);
-		return multimesh->xform_format;
-	}
-
-	_FORCE_INLINE_ bool multimesh_uses_colors(RID p_multimesh) const {
-		MultiMesh *multimesh = multimesh_owner.get_or_null(p_multimesh);
-		return multimesh->uses_colors;
-	}
-
-	_FORCE_INLINE_ bool multimesh_uses_custom_data(RID p_multimesh) const {
-		MultiMesh *multimesh = multimesh_owner.get_or_null(p_multimesh);
-		return multimesh->uses_custom_data;
-	}
-
-	_FORCE_INLINE_ uint32_t multimesh_get_instances_to_draw(RID p_multimesh) const {
-		MultiMesh *multimesh = multimesh_owner.get_or_null(p_multimesh);
-		if (multimesh->visible_instances >= 0) {
-			return multimesh->visible_instances;
-		}
-		return multimesh->instances;
-	}
-
-	_FORCE_INLINE_ RID multimesh_get_3d_uniform_set(RID p_multimesh, RID p_shader, uint32_t p_set) const {
-		MultiMesh *multimesh = multimesh_owner.get_or_null(p_multimesh);
-		if (!multimesh->uniform_set_3d.is_valid()) {
-			Vector<RD::Uniform> uniforms;
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 0;
-			u.append_id(multimesh->buffer);
-			uniforms.push_back(u);
-			multimesh->uniform_set_3d = RD::get_singleton()->uniform_set_create(uniforms, p_shader, p_set);
-		}
-
-		return multimesh->uniform_set_3d;
-	}
-
-	_FORCE_INLINE_ RID multimesh_get_2d_uniform_set(RID p_multimesh, RID p_shader, uint32_t p_set) const {
-		MultiMesh *multimesh = multimesh_owner.get_or_null(p_multimesh);
-		if (!multimesh->uniform_set_2d.is_valid()) {
-			Vector<RD::Uniform> uniforms;
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 0;
-			u.append_id(multimesh->buffer);
-			uniforms.push_back(u);
-			multimesh->uniform_set_2d = RD::get_singleton()->uniform_set_create(uniforms, p_shader, p_set);
-		}
-
-		return multimesh->uniform_set_2d;
-	}
-
-	/* SKELETON API */
-
-	RID skeleton_allocate();
-	void skeleton_initialize(RID p_skeleton);
-
-	void skeleton_allocate_data(RID p_skeleton, int p_bones, bool p_2d_skeleton = false);
-	void skeleton_set_base_transform_2d(RID p_skeleton, const Transform2D &p_base_transform);
-	void skeleton_set_world_transform(RID p_skeleton, bool p_enable, const Transform3D &p_world_transform);
-	int skeleton_get_bone_count(RID p_skeleton) const;
-	void skeleton_bone_set_transform(RID p_skeleton, int p_bone, const Transform3D &p_transform);
-	Transform3D skeleton_bone_get_transform(RID p_skeleton, int p_bone) const;
-	void skeleton_bone_set_transform_2d(RID p_skeleton, int p_bone, const Transform2D &p_transform);
-	Transform2D skeleton_bone_get_transform_2d(RID p_skeleton, int p_bone) const;
-
-	_FORCE_INLINE_ bool skeleton_is_valid(RID p_skeleton) {
-		return skeleton_owner.get_or_null(p_skeleton) != nullptr;
-	}
-
-	_FORCE_INLINE_ RID skeleton_get_3d_uniform_set(RID p_skeleton, RID p_shader, uint32_t p_set) const {
-		Skeleton *skeleton = skeleton_owner.get_or_null(p_skeleton);
-		ERR_FAIL_COND_V(!skeleton, RID());
-		ERR_FAIL_COND_V(skeleton->size == 0, RID());
-		if (skeleton->use_2d) {
-			return RID();
-		}
-		if (!skeleton->uniform_set_3d.is_valid()) {
-			Vector<RD::Uniform> uniforms;
-			RD::Uniform u;
-			u.uniform_type = RD::UNIFORM_TYPE_STORAGE_BUFFER;
-			u.binding = 0;
-			u.append_id(skeleton->buffer);
-			uniforms.push_back(u);
-			skeleton->uniform_set_3d = RD::get_singleton()->uniform_set_create(uniforms, p_shader, p_set);
-		}
-
-		return skeleton->uniform_set_3d;
-	}
 	/* Light API */
 
 	void _light_initialize(RID p_rid, RS::LightType p_type);
@@ -1498,7 +891,6 @@ public:
 	float reflection_probe_get_ambient_color_energy(RID p_probe) const;
 
 	void base_update_dependency(RID p_base, DependencyTracker *p_instance);
-	void skeleton_update_dependency(RID p_skeleton, DependencyTracker *p_instance);
 
 	/* VOXEL GI API */
 
@@ -1828,8 +1220,6 @@ public:
 	virtual uint64_t get_captured_timestamp_gpu_time(uint32_t p_index) const;
 	virtual uint64_t get_captured_timestamp_cpu_time(uint32_t p_index) const;
 	virtual String get_captured_timestamp_name(uint32_t p_index) const;
-
-	RID get_default_rd_storage_buffer() { return default_rd_storage_buffer; }
 
 	static RendererStorageRD *base_singleton;
